move watch styles to their own file
